+++ silo-core/contracts/lib/Actions.sol
@@ -225,14 +225,18 @@ library Actions {
         siloConfig.turnOnReentrancyProtection();
         siloConfig.accrueInterestForBothSilos();
 
-        (address protectedShareToken, address collateralShareToken,) = siloConfig.getShareTokens(address(this));
+        (
+            ISiloConfig.DepositConfig memory depositConfig,
+            ISiloConfig.ConfigData memory collateralConfig,
+            ISiloConfig.ConfigData memory debtConfig
+        ) = siloConfig.getConfigsForWithdraw(address(this), _args.owner);
 
         uint256 shares;
 
         // transition collateral withdraw
         address shareTokenFrom = _args.transitionFrom == ISilo.CollateralType.Collateral
-            ? collateralShareToken
-            : protectedShareToken;
+            ? depositConfig.collateralShareToken
+            : depositConfig.protectedShareToken;
 
         (assets, shares) = SiloERC4626Lib.withdraw({
             _asset: address(0), // empty token because we don't want to transfer
@@ -250,8 +254,8 @@ library Actions {
         // transition collateral deposit
         (ISilo.CollateralType depositType, address shareTokenTo) =
             _args.transitionFrom == ISilo.CollateralType.Collateral
-                ? (ISilo.CollateralType.Protected, protectedShareToken)
-                : (ISilo.CollateralType.Collateral, collateralShareToken);
+                ? (ISilo.CollateralType.Protected, depositConfig.protectedShareToken)
+                : (ISilo.CollateralType.Collateral, depositConfig.collateralShareToken);
 
         (assets, toShares) = SiloERC4626Lib.deposit({
             _token: address(0), // empty token because we don't want to transfer
@@ -263,13 +267,10 @@ library Actions {
             _collateralType: depositType
         });
 
-        // solvency check
-        ISiloConfig.ConfigData memory collateralConfig;
-        ISiloConfig.ConfigData memory debtConfig;
-
-        (collateralConfig, debtConfig) = siloConfig.getConfigsForSolvency(_args.owner);
-
-        _checkSolvencyWithoutAccruingInterest(collateralConfig, debtConfig, _args.owner);
+        // If deposit is collateral, then check the solvency.
+        if (depositConfig.silo == collateralConfig.silo) {
+            _checkSolvencyWithoutAccruingInterest(collateralConfig, debtConfig, _args.owner);
+        }
 
         siloConfig.turnOffReentrancyProtection();
 
@@ -457,17 +458,17 @@ library Actions {
 
     // this method expect interest to be already accrued
     function _checkSolvencyWithoutAccruingInterest(
-        ISiloConfig.ConfigData memory collateralConfig,
-        ISiloConfig.ConfigData memory debtConfig,
+        ISiloConfig.ConfigData memory _collateralConfig,
+        ISiloConfig.ConfigData memory _debtConfig,
         address _user
     ) private {
-        if (debtConfig.silo != collateralConfig.silo) {
-            collateralConfig.callSolvencyOracleBeforeQuote();
-            debtConfig.callSolvencyOracleBeforeQuote();
+        if (_debtConfig.silo != _collateralConfig.silo) {
+            _collateralConfig.callSolvencyOracleBeforeQuote();
+            _debtConfig.callSolvencyOracleBeforeQuote();
         }
 
         bool userIsSolvent = SiloSolvencyLib.isSolvent(
-            collateralConfig, debtConfig, _user, ISilo.AccrueInterestInMemory.No
+            _collateralConfig, _debtConfig, _user, ISilo.AccrueInterestInMemory.No
         );
 
         require(userIsSolvent, ISilo.NotSolvent());