+++ silo-core/test/foundry/Silo/borrow/Borrow.i.sol
@@ -33,7 +33,7 @@ contract BorrowIntegrationTest is SiloLittleHelper, Test {
     forge test -vv --ffi --mt test_borrow_all_zeros
     */
     function test_borrow_all_zeros() public {
-        vm.expectRevert(ISilo.InputZeroAssetsOrShares.selector);
+        vm.expectRevert(ISilo.InputZeroShares.selector);
         silo0.borrow(0, address(0), address(0));
     }
 
@@ -44,7 +44,7 @@ contract BorrowIntegrationTest is SiloLittleHelper, Test {
         uint256 assets = 0;
         address borrower = address(1);
 
-        vm.expectRevert(ISilo.InputZeroAssetsOrShares.selector);
+        vm.expectRevert(ISilo.InputZeroShares.selector);
         silo0.borrow(assets, borrower, borrower);
     }