+++ silo-core/test/foundry/Silo/preview/PreviewDeposit.i.sol
@@ -103,7 +103,7 @@ contract PreviewDepositTest is SiloLittleHelper, Test {
 
         if (previewShares1 == 0) {
             // if preview is zero for `_assets`, then deposit should also reverts
-            _depositForBorrowRevert(_assets, depositor, cType, ISilo.InputZeroAssetsOrShares.selector);
+            _depositForBorrowRevert(_assets, depositor, cType, ISilo.InputZeroShares.selector);
         } else {
             assertEq(
                 previewShares1,
@@ -132,7 +132,7 @@ contract PreviewDepositTest is SiloLittleHelper, Test {
         emit log_named_uint("previewShares1", previewShares1);
 
         if (previewShares1 == 0) {
-            _depositForBorrowRevert(_assets, depositor, cType, ISilo.InputZeroAssetsOrShares.selector);
+            _depositForBorrowRevert(_assets, depositor, cType, ISilo.InputZeroShares.selector);
         } else {
             assertEq(
                 previewShares1,