+++ silo-core/test/foundry/Silo/repay/Repay.i.sol
@@ -26,7 +26,7 @@ contract RepayTest is SiloLittleHelper, Test {
     forge test -vv --ffi --mt test_repay_zeros
     */
     function test_repay_zeros() public {
-        vm.expectRevert(ISilo.InputZeroAssetsOrShares.selector);
+        vm.expectRevert(ISilo.InputZeroShares.selector);
         silo0.repay(0, address(0));
     }
 
@@ -67,7 +67,7 @@ contract RepayTest is SiloLittleHelper, Test {
         _createDebt(assets, borrower);
         vm.warp(block.timestamp + 50 * 365 days); // interest must be big, so conversion 1 asset => share be 0
 
-        vm.expectRevert(ISilo.ReturnZeroAssetsOrShares.selector);
+        vm.expectRevert(ISilo.ReturnZeroShares.selector);
         silo1.repay(assets, borrower);
     }